Load health icons from a CSV file.
Health icons are specified in 'assets/preload/data/healthIcons.csv'.

The second column in the file determines the image the icon will be
loaded from. The image must be a grid of 150x150 icons. All base-game
icons are loaded from 'iconGrid'.

The third and fourth columns determine which icons are used when the
character is alive or dead, resectively. These are indexes into the
image specified in the second column, from left to right, top to bottom.

Adding 'noaa' in the fifth column will disable antialiasing, which is
used for pixel icons.