Implement ImGui rendering using SDL and OpenGL 3
Input still doesn't work since SDL events are not available to Haxe code
and there is no way to obtain them without modifying Lime. However,
rendering is completely functional in this commit.

This is accomplished by adding a RenderEvent listener to an OpenFL
sprite. These events are fired at the perfect time for OpenGL calls to
be called without interfering with HaxeFlixel's rendering. Note that the
event listener must be attached to a sprite and not the stage or else
this won't work.