Stop SDL from replacing main with SDL_main
This is done by default on Windows so SDL can provide a WinMain function
that forwards to the user-provided main function. Applications using
main show a console window and applications using WinMain do not.

This is not needed here since Lime already provides a WinMain function,
and this macro breaks Haxe functions called main. This commit adds a
line that tells SDL not to define this macro.