Update build instructions to fix broken game
When using the default setting of dynamic linking, Lime uses OpenAL Soft
as its audio library, but when statically linking, Lime uses MojoAL.
MojoAL happens to have an issue with retrieving the current audio
position that breaks the entire game.

This commit revises the build instructions with a command that forces
Lime to use OpenAL even when statically linking.