Start transitioning to new chart format
This commit adds code to convert charts from the old format to the new
format just-in-time upon loading the song. However, these new-format
fields are still unused since the game has not been updated to use them.

The next step is to refactor the rest of the game's code to use the new
note track system and then remove the old fields ('info' and 'chart').