Create Keybinding class and add gamepad support
The Keybinding class is used to check the state of a keybinding. Unlike
HaxeFlixel actions, you do not need to create separate keybindings to
detect different states like justPressed and pressed.

An example of using the Keybinding class is as follows:

    private final doAThing = new Keybinding("a-mod:do-a-thing").load();

    override public function update(elapsed:Float)
    {
        // ...
        if (doAThing.justPressed)
        {
            // Do the thing.
        }
    }

Keybindings are currently never used anywhere in the codebase, but the
next step is to integrate keybindings and get rid of the Controls class.